feat: first basic implementation of a bitvector as primitive type
The `BitVector` type represents just a stream of binary values.
It can be used to represent numbers or do other fancy things with it.

Under the hood it uses `std::vector<u8>`, as `std::vector<bool>` is to
be avoided. The implementation can change in the future to something
like an array of integer.